# Api4ApiKeys/MinimalAPI/swagger.json
{
  "x-generator": "NSwag v14.6.3.0 (NJsonSchema v11.5.2.0 (Newtonsoft.Json v13.0.0.0))",
  "openapi": "3.0.0",
  "info": {
    "title": "APIKeyAPI v1",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5141"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "operationId": "Get",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/key/id/{id}": {
      "get": {
        "operationId": "GetKeyId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "put": {
        "operationId": "PutKeyId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "inputKey",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SecretKey"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "operationId": "DeleteKeyId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/key/name/{keyName}": {
      "get": {
        "operationId": "GetKeyName",
        "parameters": [
          {
            "name": "keyName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "put": {
        "operationId": "PutKeyName",
        "parameters": [
          {
            "name": "keyName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "inputKey",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SecretKey"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/key": {
      "post": {
        "operationId": "PostKey",
        "requestBody": {
          "x-name": "inputKey",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SecretKey"
              }
            }
          },
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "SecretKey": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "keyName": {
            "type": "string",
            "maxLength": 100
          },
          "secretValue": {
            "type": "string",
            "maxLength": 100
          }
        }
      }
    }
  }
}